DOCAPI

+++ clientIP/src/component/Nav.jsx
@@ -1,17 +1,35 @@
 import { Link, useNavigate } from "react-router-dom";
-const navigate = useNavigate;
+import Swal from "sweetalert2";
+
+
 export default function Nav() {
   let navigate = useNavigate();
   const handleLogut = (e) => {
     e.preventDefault();
-    localStorage.clear();
-    navigate("/login");
+    Swal.fire({
+      title: "Logout?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, Logout!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "Succes Logout!",
+         
+          icon: "success",
+        });
+        localStorage.clear();
+        navigate("/login");
+      }
+    });
   };
   return (
     <>
       <header className="bg-white">
         <div className="mx-auto flex h-16 max-w-screen-xl items-center gap-8 px-4 sm:px-6 lg:px-8">
-          <a className="block text-teal-600" href="#">
+          <Link className="block text-teal-600" to="/">
             <span className="sr-only">Home</span>
             <svg
               className="h-8"
@@ -24,69 +42,19 @@ export default function Nav() {
                 fill="currentColor"
               />
             </svg>
-          </a>
+          </Link>
 
           <div className="flex flex-1 items-center justify-end md:justify-between">
             <nav aria-label="Global" className="hidden md:block">
               <ul className="flex items-center gap-6 text-sm">
                 <li>
                   <Link
-                    className="text-gray-500 transition hover:text-gray-500/75"
+                    className="text-black-500 transition hover:text-black-500/75"
                     to={"/plants"}
                   >
                     {" "}
                     Add Plant{" "}
                   </Link>
-                </li>
-
-                <li>
-                  <a
-                    className="text-gray-500 transition hover:text-gray-500/75"
-                    href="#"
-                  >
-                    {" "}
-                    Careers{" "}
-                  </a>
-                </li>
-
-                <li>
-                  <a
-                    className="text-gray-500 transition hover:text-gray-500/75"
-                    href="#"
-                  >
-                    {" "}
-                    History{" "}
-                  </a>
-                </li>
-
-                <li>
-                  <a
-                    className="text-gray-500 transition hover:text-gray-500/75"
-                    href="#"
-                  >
-                    {" "}
-                    Services{" "}
-                  </a>
-                </li>
-
-                <li>
-                  <a
-                    className="text-gray-500 transition hover:text-gray-500/75"
-                    href="#"
-                  >
-                    {" "}
-                    Projects{" "}
-                  </a>
-                </li>
-
-                <li>
-                  <a
-                    className="text-gray-500 transition hover:text-gray-500/75"
-                    href="#"
-                  >
-                    {" "}
-                    Blog{" "}
-                  </a>
                 </li>
               </ul>
             </nav>
@@ -95,7 +63,7 @@ export default function Nav() {
               <div className="sm:flex sm:gap-4">
                 <Link
                   className="block rounded-md bg-teal-600 px-5 py-2.5 text-sm font-medium text-white transition hover:bg-teal-700"
-                  to={"/"}
+                  to={"/API"}
                 >
                   Select Location
                 </Link>
@@ -107,8 +75,6 @@ export default function Nav() {
                   LOGOUT
                 </button>
               </div>
-
-             
             </div>
           </div>
         </div>
